Support compiling with GOOS=js GOARCH=wasm
`GOOS=js GOARCH=wasm make build` should work now.
