+++ cmd/community/community.go
@@ -5,7 +5,6 @@ import (
 )
 
 func init() {
-	CommunityCmd.AddCommand(CreateManifestCmd)
 	CommunityCmd.AddCommand(ListIconsCmd)
 	CommunityCmd.AddCommand(LoadAppCmd)
 	CommunityCmd.AddCommand(SpellCheckCmd)
