+++ cmd/community/manifestprompt.go
@@ -1,3 +1,5 @@
+//go:build !js && !wasm
+
 package community
 
 import (
