+++ cmd/create.go
@@ -1,3 +1,5 @@
+//go:build !js && !wasm
+
 package cmd
 
 import (
