+++ main_nonjs.go
@@ -1,0 +1,11 @@
+//go:build !js && !wasm
+
+package main
+
+import (
+	"tidbyt.dev/pixlet/cmd"
+)
+
+func init() {
+	rootCmd.AddCommand(cmd.CreateCmd)
+}
