+++ render/image_webp_js.go
@@ -1,0 +1,11 @@
+//go:build js && wasm
+
+package render
+
+import (
+	"fmt"
+)
+
+func (p *Image) InitFromWebP(data []byte) error {
+	return fmt.Errorf("WebP not supported in WASM")
+}
