+++ server/watcher/fs_js.go
@@ -1,0 +1,11 @@
+//go:build js && wasm
+
+package watcher
+
+import (
+	"fmt"
+)
+
+func (w *Watcher) Run() error {
+	return fmt.Errorf("file watching not supported in WASM")
+}
